refactor(data-model): rename single-function json-wire files to match exports
Apply the generalized single-export file-naming rule (file base = export
name) to the two single-function files in `src/json-wire/`:

- `create-default-registry.ts` -> `createDefaultRegistry.ts`
- `make-problematic.ts` -> `makeProblematic.ts`

Pure renames plus import-line repoints (the barrel `index.ts`, the
6 handler files that import `makeProblematic`, and `JsonEncodingContext.ts`
which imports `createDefaultRegistry`); file bodies are byte-identical. The
multi-export files `json-wire-types.ts` and `json-encoding.ts` stay kebab
(the rule applies to single-export files only).
